+++ src/test/kotlin/ru/yandex/money/gradle/plugins/javapublishing/PublishingTest.kt
@@ -14,6 +14,8 @@ class PublishingTest : AbstractReleaseTest() {
         javaArtifactPublishSettings {
             artifactId = "test_artifact_id"
             groupId = "test_group_id"
+            snapshotRepository = "https://nexus.yamoney.ru/repository/snapshots/"
+            releaseRepository = "https://nexus.yamoney.ru/repository/release/"
         }
         """)
         runTasksSuccessfully("build", "pTML", "--info")