+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Добавлены настройки snapshotRepository и releaseRepository - для указания адресов репозиториев.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [2.2.0](https://bitbucket.yamoney.ru/projects/BACKEND-GRADLE-PLUGINS/repos/java-artifact-publish-plugin/pull-requests/1) (03-07-2020)